Ingestion-Lambda/requirements.txt
snowflake-connector-python==3.1.0
python-dotenv
boto3==1.28.30